backend/requirements.txt
fastapi[all]
httptools==0.6.1
opencv-python==4.7.0.72
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.26.0
rest-framework-simplejwt==0.0.1
scikit-image==0.22.0
scikit-learn==1.3.2
torch==2.4.0
torchvision==0.19.0
transformers==4.43.3
uvicorn==0.30.1
gradio_client